# frontend/altairis-frontend/src/app/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Altairis Backoffice API",
    "description": "B2B hotel distribution backoffice API",
    "version": "v1"
  },
  "paths": {
    "/api/hotels/{hotelId}/availability": {
      "get": {
        "tags": [
          "Availability"
        ],
        "summary": "Get hotel inventory availability grid for a date range.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelInventoryGridDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get dashboard summary metrics.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "List hotels with pagination.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDtoPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hotels"
        ],
        "summary": "Create a new hotel.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels/{id}": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Get a single hotel by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels/search": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Search hotels by name, city or country.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "List reservations with optional filters and pagination.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "enum": [
                "Pending",
                "Confirmed",
                "Cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDtoPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reservations"
        ],
        "summary": "Create a new reservation.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/{id}": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "Get a single reservation by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/{id}/status": {
      "patch": {
        "tags": [
          "Reservations"
        ],
        "summary": "Update reservation status (confirm or cancel).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationStatusDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationStatusDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/room-categories": {
      "get": {
        "tags": [
          "RoomCategories"
        ],
        "summary": "List all room categories.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomCategoryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RoomCategories"
        ],
        "summary": "Create a new room category.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/hotels/{hotelId}/room-types": {
      "get": {
        "tags": [
          "RoomTypes"
        ],
        "summary": "List room types for a hotel.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomTypeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RoomTypes"
        ],
        "summary": "Create a room type for a hotel.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomTypeDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateHotelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "roomTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "checkIn": {
            "type": "string",
            "format": "date-time"
          },
          "checkOut": {
            "type": "string",
            "format": "date-time"
          },
          "guestName": {
            "type": "string",
            "nullable": true
          },
          "roomsBooked": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateRoomCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateRoomTypeDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DashboardDto": {
        "type": "object",
        "properties": {
          "totalHotels": {
            "type": "integer",
            "format": "int32"
          },
          "totalReservations": {
            "type": "integer",
            "format": "int32"
          },
          "availableRoomsToday": {
            "type": "integer",
            "format": "int32"
          },
          "occupancyPercentage": {
            "type": "number",
            "format": "double"
          },
          "recentReservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationDto"
            },
            "nullable": true
          },
          "occupancyByHotel": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelOccupancyDto"
            },
            "nullable": true
          },
          "reservationsByStatus": {
            "$ref": "#/components/schemas/ReservationsByStatusDto"
          }
        },
        "additionalProperties": false
      },
      "HotelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "HotelDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelDto"
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HotelInventoryGridDto": {
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "integer",
            "format": "int32"
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "roomTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomTypeInventoryRow"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HotelOccupancyDto": {
        "type": "object",
        "properties": {
          "hotelName": {
            "type": "string",
            "nullable": true
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          },
          "bookedRooms": {
            "type": "integer",
            "format": "int32"
          },
          "occupancyPercentage": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "InventoryCell": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          },
          "availableRooms": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReservationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roomTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "checkIn": {
            "type": "string",
            "format": "date-time"
          },
          "checkOut": {
            "type": "string",
            "format": "date-time"
          },
          "guestName": {
            "type": "string",
            "nullable": true
          },
          "roomsBooked": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "enum": [
              "Pending",
              "Confirmed",
              "Cancelled"
            ],
            "type": "string"
          },
          "roomTypeName": {
            "type": "string",
            "nullable": true
          },
          "hotelName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReservationDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationDto"
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReservationsByStatusDto": {
        "type": "object",
        "properties": {
          "pending": {
            "type": "integer",
            "format": "int32"
          },
          "confirmed": {
            "type": "integer",
            "format": "int32"
          },
          "cancelled": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RoomCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoomTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "hotelId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RoomTypeInventoryRow": {
        "type": "object",
        "properties": {
          "roomTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "roomTypeName": {
            "type": "string",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          },
          "dates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryCell"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateReservationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "Pending",
              "Confirmed",
              "Cancelled"
            ],
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}